Simplify OrthoMeshLayer
